Reorg pagetable rewrite.
